Cast the area accumulators to float32; 1024 = 32^2 names the node
CHECK 8 failed again with a BYTE-IDENTICAL error: "1 bands for 17305600
pixels = 132.0 MiB > 80.0 MiB". The two numbers in it name the node, and
this time they name it precisely rather than approximately.

  132 MiB / 17,305,600 px  = 8.0 bytes/px exactly -> a float64 band
  sqrt(17,305,600) = 4160  = 4096 + 2*32          -> a 32 px halo

A 32 px halo at the 10 m analysis scale is maxComponentSizePx 1024 = 32^2,
EXACTLY. The previous commit guessed edgeErodeM - 250 m = 25 px, rounding
UP to 32 - which fits loosely; 1024 = 32^2 fits exactly, and supersedes it.
The bands carrying connectedPixelCount's halo while being one band of
float64 each are the area accumulators, both built on ee.Image.pixelArea(),
which is a double.

WHY THIS CAST IS FREE AND THE EARLIER ONES WERE NOT ALL EQUAL. areaImg and
tAreaImg are output ACCUMULATORS, not decision inputs. Nothing downstream
compares them against a threshold, so no detection can change. And there is
no precision to lose: pixelArea at 10 m is ~100 m2 and npix is capped at
maxComponentSizePx, so the product tops out near 1.0e5, while float32 is
exact for every integer to 2^24 = 1.7e7. 4160^2 x 4 = 66.0 MiB, inside the
limit.

That is NOT true of the detector's `work`, `mu` and `thr`, which ARE the
threshold comparison. Casting those changes decisions in their last bits
and stays a deliberate, reported change rather than a quiet one. The README
now tabulates which casts are free and which are not, so the distinction
survives past this commit.

Also recorded in the README: how to read this error class at all. M/N gives
bytes per pixel, and sqrt(N) minus 4096, halved, gives the halo - which
between them identify the node. Three exports and a CHECK have died on this
limit and the error has out-diagnosed every estimate I have made about it.

And why tileScale is not the lever here: it partitions the REDUCER's
aggregation, while these failures are in the image graph, whose
neighbourhood ops tile themselves. That is why the tile stayed 4096-based
at tileScale 16.

CI 147 -> 148, mutation-tested by reverting both bands to double.

NOT VERIFIED, and worth stating plainly: the pasted error was byte-identical
to the previous one. That is consistent with the tab predating the last
merge OR with the scene-mask cast never having been the node - the second
of which this commit's arithmetic now suggests. The BUILD line settles which
on the next paste.

+++ studies/sb02-vessel-proxy/sb02_sar_export.js
@@ -941,8 +941,27 @@ function detectVessels(img, aoi, params) {
     maxSize: params.maxComponentSizePx
   }).select('labels');
 
-  var areaImg = ee.Image.pixelArea().updateMask(target).rename('area');
-  var tAreaImg = npix.multiply(ee.Image.pixelArea()).updateMask(target).rename('tarea');
+  // float32, AND THIS ONE IS FREE IN A WAY THE EARLIER TYPE FIXES WERE NOT.
+  //
+  // CHECK 8 failed with "1 bands for 17305600 pixels = 132.0 MiB > 80.0 MiB".
+  // 132 MiB / 17,305,600 px is 8.0 bytes/px - float64 - over 4160^2, which is a
+  // 4096 tile plus a 32 px halo. The halo identifies the node:
+  //   maxComponentSizePx 1024 IS 32 x 32, exactly.
+  // (An earlier commit guessed edgeErodeM, 250 m = 25 px rounding UP to 32.
+  // 1024 = 32^2 is the better fit, and it is these two bands that carry the
+  // connectedPixelCount halo while being one band of float64 each.)
+  // ee.Image.pixelArea() is a double, so both of these are doubles.
+  //
+  // WHY THIS COSTS NOTHING, unlike casting the detector's `work` image. These
+  // are output ACCUMULATORS, not decision inputs: nothing downstream compares
+  // them against a threshold, so no detection can change. And there is no
+  // precision to lose - pixelArea at 10 m is ~100 m2 and npix is capped at
+  // maxComponentSizePx, so the product tops out near 1.0e5, while float32
+  // represents every integer exactly to 2^24 = 1.7e7.
+  // 4160^2 x 4 bytes = 66.0 MiB, inside the limit.
+  var areaImg = ee.Image.pixelArea().toFloat().updateMask(target).rename('area');
+  var tAreaImg = npix.toFloat().multiply(ee.Image.pixelArea().toFloat())
+                     .updateMask(target).rename('tarea');
 
   var common = {
     geometry: aoi,
